# Documentacion/openapi.yaml
openapi: 3.0.3
info:
  title: Plants App API
  description: |
    API REST del TFG "Plants App" — aplicación móvil de gestión de plantas con
    funcionalidades de comunidad, chat en tiempo real, recordatorios y datos climáticos.
    Arquitectura: Clean Architecture (Node.js + Express + MongoDB + Socket.IO).
  version: 1.0.0
  contact:
    name: TFG — 2 DAM
servers:
  - url: http://localhost:3000/api
    description: Servidor de desarrollo local

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTES COMUNES
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido tras login o registro.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: Token inválido o expirado.
      required: [status, code, message]

    # ── Paginación ────────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        hasMore:
          type: boolean
          example: true

    # ── Auth ──────────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          example: Ana García
        email:
          type: string
          format: email
          example: ana@example.com
        password:
          type: string
          minLength: 8
          example: SecurePass1!
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: ana@example.com
        password:
          type: string
          example: SecurePass1!
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT de acceso (expira en 7 días).
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/UserPublic'

    # ── Usuario ───────────────────────────────────────────────────────────────
    UserPublic:
      type: object
      properties:
        id:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d1
        name:
          type: string
          example: Ana García
        email:
          type: string
          example: ana@example.com
        photo:
          type: string
          nullable: true
          example: https://res.cloudinary.com/demo/image/upload/avatar.jpg
        bannerPhoto:
          type: string
          nullable: true
          example: https://res.cloudinary.com/demo/image/upload/banner.jpg
        bio:
          type: string
          nullable: true
          example: Amante de las plantas tropicales.
        location:
          type: string
          nullable: true
          example: Sevilla
        locationLat:
          type: number
          format: double
          nullable: true
          example: 37.3891
        locationLon:
          type: number
          format: double
          nullable: true
          example: -5.9845
        role:
          type: string
          enum: [user, admin]
          example: user
        bannedUntil:
          type: string
          format: date-time
          nullable: true
          description: Fecha hasta la que el usuario está baneado. null si no está baneado.
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
          example: '2026-01-15T10:00:00.000Z'
    UserPreferences:
      type: object
      properties:
        appearInChatSearch:
          type: boolean
          example: true
        considerWeatherByDefault:
          type: boolean
          example: false
        isPrivate:
          type: boolean
          default: false
          description: Si true, el perfil no aparece en el feed público ni acepta nuevas conversaciones.
...
    ParticipantSummary:
      type: object
      properties:
        id:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d1
        name:
          type: string
          example: Ana García
        photo:
          type: string
          nullable: true
    Conversation:
      type: object
      properties:
        id:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d4
        participant:
          $ref: '#/components/schemas/ParticipantSummary'
        lastMessage:
          $ref: '#/components/schemas/Message'
          nullable: true
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        unreadCount:
          type: integer
          example: 3
        isParticipantDeleted:
          type: boolean
          description: True si el otro participante ha eliminado su cuenta.
        createdAt:
          type: string
          format: date-time
    Message:
      type: object
      properties:
        id:
          type: string
          example: 64f1a2b3c4d5e6f7a8b9c0d5
        conversationId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        text:
          type: string
        contentMeta:
          type: object
          nullable: true
          description: Metadatos del contenido (ej. link preview, imagen embebida).
        status:
          type: string
          enum: [pending, sent, delivered, read]
          description: "pending→sent→delivered→read. sent = servidor recibió; delivered = destinatario online."
          example: delivered
        tempId:
          type: string
          nullable: true
          description: ID temporal asignado por el cliente para correlación optimista.
        createdAt:
          type: string
          format: date-time
    SendMessageRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 2000
          example: Hola! Vi tu post sobre la monstera.
        contentMeta:
          type: object
          nullable: true
          description: Metadatos opcionales del contenido.
        tempId:
          type: string
          nullable: true
          description: ID temporal del cliente para correlación optimista.

    # ── Recordatorios ─────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: No autorizado — token ausente o inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Error de validación en el cuerpo de la petición.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitExceeded:
      description: |
        Demasiadas peticiones. Cuerpo: `{ code: 'RATE_LIMIT_EXCEEDED',
        message, details: { retryAfterMinutes } }`. Las cabeceras estándar
        `RateLimit-Limit`, `RateLimit-Remaining`, `RateLimit-Reset` y
        `Retry-After` indican cuándo se puede reintentar.
      headers:
        RateLimit-Limit:
          schema: { type: integer }
          description: Máximo de peticiones permitidas en la ventana.
        RateLimit-Remaining:
          schema: { type: integer }
          description: Peticiones restantes en la ventana actual.
        RateLimit-Reset:
          schema: { type: integer }
          description: Segundos hasta el reset de la ventana.
        Retry-After:
          schema: { type: integer }
          description: Segundos hasta poder reintentar.
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                example: RATE_LIMIT_EXCEEDED
              message:
                type: string
                example: Demasiados intentos, espera unos minutos.
              details:
                type: object
                properties:
                  retryAfterMinutes:
                    type: integer
                    example: 14

# ─────────────────────────────────────────────────────────────────────────────
# SEGURIDAD GLOBAL
# ─────────────────────────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrar nuevo usuario
      description: |
        Crea una cuenta de usuario y devuelve un JWT de acceso.
        **Rate limit:** 10 peticiones / 15 min por IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email ya registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      description: |
        Autentica al usuario con email y contraseña y devuelve un JWT.
        **Rate limit:** 10 peticiones / 15 min por IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login correcto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /auth/validate-token:
    get:
      tags: [Auth]
      summary: Validar token JWT
      description: Verifica que el token Bearer es válido y devuelve el usuario autenticado.
      responses:
        '200':
          description: Token válido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar token JWT
      description: |
        Genera un token JWT nuevo con expiración fresca (`JWT_EXPIRES_IN`,
        30d por defecto) para el usuario autenticado. F7-056 / decisión #175.

        El frontend lo invoca al arrancar la app si al token actual le quedan
        menos de 7 días para expirar (sesión persistente híbrida — sin
        refresh tokens completos).

        No requiere body: el `userId` se extrae del JWT actual a través del
        `AuthMiddleware`. Idempotente — se puede llamar varias veces seguidas.
      responses:
        '200':
          description: Token renovado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: El usuario fue soft-deleted entre la emisión del token y este refresh.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'

  # ════════════════════════════════════════════════════════════════════════════
  # PLANTS
  # ════════════════════════════════════════════════════════════════════════════
...
  /chat:
    get:
      tags: [Chat]
      summary: Listar conversaciones del usuario
      description: Devuelve todas las conversaciones del usuario autenticado, ordenadas por actividad reciente.
      responses:
        '200':
          description: Lista de conversaciones.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conversation'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Chat]
      summary: Obtener o crear conversación con un usuario
      description: Devuelve la conversación existente con el otro usuario o la crea si no existe (get-or-create).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otherUserId]
              properties:
                otherUserId:
                  type: string
                  example: 64f1a2b3c4d5e6f7a8b9c0d1
      responses:
        '200':
          description: Conversación existente o recién creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chat/{id}/messages:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID de la conversación.
    get:
      tags: [Chat]
      summary: Obtener mensajes de una conversación
      description: Devuelve mensajes paginados (cursor-based), del más reciente al más antiguo.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Mensajes paginados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Chat]
      summary: Enviar mensaje
      description: Envía un nuevo mensaje en la conversación. El servidor emite `message:new` por Socket.IO.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Mensaje enviado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /chat/{id}/read:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: ID de la conversación.
    post:
      tags: [Chat]
      summary: Marcar mensajes como leídos
      description: Marca todos los mensajes no leídos de la conversación como leídos. Emite `messages:read` por Socket.IO.
      responses:
        '204':
          description: Mensajes marcados como leídos.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════════════════════════
  # USERS
  # ════════════════════════════════════════════════════════════════════════════
...